creamos el repositorio

+++ src/main/java/com/v4/v4/productoRepository.java
@@ -5,7 +5,7 @@ import com.v4.v4.producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface productoRepository extends JpaRepository<producto, Integer> {
+public interface productoRepository extends JpaRepository<producto, Long> {
     List<producto> findByNombreConiningIgnoreCase(String nombre);
     List<producto> findBycantidadlessThanEqual(int cantidad);
